NPC Map Show
*Changed ini entry from Genral Settings.ini to Character.ini
*Made the made update in 1 min intervals to fix issue with respawns not showing up.

After much thought & the fact that respawns were not showing up the code has been changed.  Too save on CPU cycles the ini line was move to the character ini to turn the feature off on characters that you do not wish it to be used on.

new character.ini line:

[Misc]
MapNPC (On/Off)=Off